requirements.txt
streamlit
google-generativeai
speechrecognition
gtts
pyaudio